# resources/Wasteless_API_Spec.yml
openapi: 3.0.0
servers:
  - description: SwaggerHub API Auto Mocking
    url: >-
      https://virtserver.swaggerhub.com/matthewminish/seng302-2021-api-spec/3.0.1
info:
  description: API specification for the SENG302 2021 project team 800
  version: 6.0.0
  title: Wasteless
  contact:
    email: seng302@canterbury.ac.nz
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
...
security:
  - CookieAuth:
      - user
      - globalApplicationAdmin
      - defaultGlobalApplicationAdmin
paths:
...
  /businesses:
    post:
      tags:
        - 'U5: Creating Business Accounts'
        - Business Accounts
      summary: Create a new business entity
      security:
        - CookieAuth:
            - user
            - globalApplicationAdmin
            - defaultGlobalApplicationAdmin
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Business'
        description: >-
          New business account data, id will be generated automatically by the
          server
      responses:
        '201':
          description: Business account created successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  businessId:
                    type: integer
                    description: The ID of the business that has just been created
                    example: 100
        '401':
          $ref: '#/components/responses/UnauthorizedError'
  /businesses/{id}:
    get:
      tags:
        - 'U5: Creating Business Accounts'
        - Business Accounts
      summary: Retrieve a specific business account
      security:
        - CookieAuth:
            - user
            - globalApplicationAdmin
            - defaultGlobalApplicationAdmin
      parameters:
        - name: id
          in: path
          description: Business ID
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Business retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Business'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '406':
          $ref: '#/components/responses/NotAcceptableError'
  /businesses/{id}/makeAdministrator:
    put:
      tags:
        - 'U5: Creating Business Accounts'
        - Business Accounts
      summary: Add an individual as an administrator for a business
      security:
        - CookieAuth:
            - user
            - globalApplicationAdmin
            - defaultGlobalApplicationAdmin
      parameters:
        - name: id
          in: path
          description: Business ID
          required: true
          schema:
            type: integer
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                userId:
                  type: integer
                  description: The ID of the user to add as a business administrator
                  example: 100
      responses:
        '200':
          description: Individual added as an administrator successfully
        '400':
          description: >-
            The individual could not be added as an administrator, perhaps
            because it does not exist or is already an administrator.
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          description: >-
            Forbidden: Returned when the user making the request is not the
            primary administrator of the business AND is not a global
            application admin.
        '406':
          $ref: '#/components/responses/NotAcceptableError'
  /businesses/{id}/removeAdministrator:
    put:
      tags:
        - 'U5: Creating Business Accounts'
        - Business Accounts
      summary: Remove an individual from administering a business
      security:
        - CookieAuth:
            - user
            - globalApplicationAdmin
            - defaultGlobalApplicationAdmin
      parameters:
        - name: id
          in: path
          description: Business ID
          required: true
          schema:
            type: integer
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                userId:
                  type: integer
                  description: The ID of the user to remove as a business administrator
                  example: 100
      responses:
        '200':
          description: Inidividual removed from administering successfully
        '400':
          description: >-
            The individual could not be removed from administering the business,
            perhaps because it does not exist, was not an administrator of the
            business, or is the primary administrator of the business.
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          description: >-
            Forbidden: Returned when the user making the request is not the
            primary administrator of the business AND is not a global
            application admin.
        '406':
          $ref: '#/components/responses/NotAcceptableError'
...
components:
  securitySchemes:
    CookieAuth:
      type: apiKey
      in: cookie
      name: JSESSIONID
  responses:
...
    UnauthorizedError:
      description: Access token is missing or invalid
...
    NotAcceptableError:
      description: >-
        The requested route does exist (so not a 404) but some part of the
        request is not acceptable, for example trying to access a resource by an
        ID that does not exist.
...
  schemas:
...
    Business:
      type: object
      description: Business account entity
      required:
        - name
        - address
        - businessType
      properties:
        id:
          type: integer
          example: 100
          readOnly: true
          description: Automatically generated and assigned by the server
        administrators:
          type: array
          items:
            $ref: '#/components/schemas/User'
          readOnly: true
        primaryAdministratorId:
          type: integer
          example: 20
          description: >-
            The ID of the 'primary' administrator of the business. This is
            automatically given to the creator of the business.
        name:
          type: string
          example: Lumbridge General Store
        description:
          type: string
          example: A one-stop shop for all your adventuring needs
        address:
          $ref: '#/components/schemas/Address'
        businessType:
          type: string
          enum:
            - Accommodation and Food Services
            - Retail Trade
            - Charitable organisation
            - Non-profit organisation
        created:
          type: string
          format: date-time
          example: '2020-07-14T14:52:00Z'
          readOnly: true
          description: Date and time when the account was created
    Address:
      type: object
      description: Entity representing some physical address
      required:
        - country
      properties:
        streetNumber:
          type: string
          example: 3/24
          writeOnly: true
        streetName:
          type: string
          example: Ilam Road
          writeOnly: true
        suburb:
          type: string
          example: Upper Riccarton
          writeOnly: true
        city:
          type: string
          example: Christchurch
        region:
          type: string
          example: Canterbury
        country:
          type: string
          example: New Zealand
        postcode:
          type: string
          example: '90210'
          writeOnly: true
    User:
      type: object
      description: Basic user entity
      required:
        - firstName
        - lastName
        - email
        - dateOfBirth
        - homeAddress
        - password
      properties:
        id:
          type: integer
          example: 100
          readOnly: true
          description: Automatically generated and assigned by the server
        firstName:
          type: string
          example: John
        lastName:
          type: string
          example: Smith
        middleName:
          type: string
          example: Hector
        nickname:
          type: string
          example: Jonny
        bio:
          type: string
          example: Likes long walks on the beach
        email:
          type: string
          format: email
          example: johnsmith99@gmail.com
          writeOnly: true
        primaryImageId:
          type: integer
          example: 1
          description: >-
            The ID of the 'primary' image chosen by the user. By default it is
            the first image the use uploads.
        countryForCurrency:
          type: string
          description: >-
            The country in which the currency is used, if not initialized then
            default to New Zealand
          example: New Zealand
        dateOfBirth:
          type: string
          format: date
          example: '1999-04-27'
          writeOnly: true
        phoneNumber:
          type: string
          example: +64 3 555 0129
          writeOnly: true
        homeAddress:
          $ref: '#/components/schemas/Address'
        password:
          type: string
          example: 1337-H%nt3r2
          format: password
          writeOnly: true
          description: Should not be stored in plain text (i.e hash it)
        created:
          type: string
          format: date-time
          example: '2020-07-14T14:32:00Z'
          readOnly: true
          description: Date and time when the account was created
        role:
          type: string
          readOnly: true
          description: >-
            (Implemented in U4): This property should only be shown to Global
            Application Admins
          enum:
            - user
            - globalApplicationAdmin
            - defaultGlobalApplicationAdmin
        businessesAdministered:
          type: array
          readOnly: true
          description: >-
            (Implemented in U6): A list of businesses for which the user is an
            administrator
          items:
            $ref: '#/components/schemas/Business'
...